WIP: Adding async/await to the entire project

+++ server/server.js
@@ -1,7 +1,9 @@
 require('./config/config');
+const asyncHandler = require('express-async-handler');
 const express = require('express');
 const _ = require('lodash');
 const { authenticate } = require('./middleware/authenticate');
+console.error('A unicorn got in');
 
 const { ObjectID } = require('mongodb');
 const { mongoose } = require('./db/mongoose');
@@ -12,56 +14,41 @@ const bcrypt = require('bcryptjs');
 const app = express();
 const port = process.env.PORT;
 
-const showSomeshit = someshit => console.log(someshit);
-
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 
 app
-  .post('/todos', authenticate, (req, res) => {
-    let todo = new Todo({
+  .post('/todos', authenticate, asyncHandler(async (req, res) => {
+    const todo = new Todo({
       text: req.body.text,
       _creator: req.user._id
     });
-    todo.save()
-      .then(
-        doc => res.status(201).send(doc))
-      .catch(e => res.status(400).send(e.message));
-  })
-  .get('/todos', authenticate, (req, res) => {
-    Todo.find({
-        _creator: req.user._id
-      })
-      .then(
-        todos => res.status(200).send({ todos }))
-      .catch(e => res.status(400).send());
-  })
-  .get('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
-    if (!ObjectID.isValid(id)) return res.status(404).json({ error: 'Invalid id' });
-    Todo.findOne({
-        _id: id,
-        _creator: req.user._id
-      })
-      .then(todo => {
-        if (!todo) throw new Error('Todo not found');
-        res.status(200).json({ todo });
-      }).catch(e => res.status(400).send(e.message));
-  })
-  .delete('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
-    if (!ObjectID.isValid(id)) return res.status(404).json({ error: 'Invalid id' });
-    Todo.findOneAndRemove({
-      _id: id,
-      _creator: req.user._id
-    }).then(todo => {
-      if (!todo) return res.status(404).json({ message: 'Todo not found' });
-      res.status(200).json({ todo, message: 'Todo was successfully removed' });
-    }).catch(e => res.status(400).send());
-  })
-  .patch('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
 
+    await todo.save();
+    res.status(201).send(todo);
+  }))
+  .get('/todos', authenticate, asyncHandler(async (req, res) => {
+    const todos = await Todo.find({ _creator: req.user._id });
+    res.status(200).send({ todos });
+  }))
+  .get('/todos/:id', authenticate, asyncHandler(async (req, res) => {
+    let id = req.params.id;
+    if (!ObjectID.isValid(id)) return res.status(404).json({ error: 'Invalid id' });
+
+    const todo = await Todo.findOne({ _id: id, _creator: req.user._id });
+    if (!todo) throw new Error('Todo not found.');
+    res.status(200).json({ todo });
+  }))
+  .delete('/todos/:id', authenticate, asyncHandler(async (req, res) => {
+    let id = req.params.id;
+    if (!ObjectID.isValid(id)) return res.status(404).json({ error: 'Invalid id' });
+
+    const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
+    if (!todo) return res.status(404).json({ message: 'Todo not found' });
+    res.status(200).json({ todo, message: 'Todo was successfully removed' });
+  }))
+  .patch('/todos/:id', authenticate, asyncHandler(async (req, res) => {
+    let id = req.params.id;
     if (!ObjectID.isValid(id)) return res.status(404).json({ error: 'Invalid id' });
 
     let body = _.pick(req.body, ['text', 'completed']);
@@ -73,42 +60,42 @@ app
       body.completedAt = null;
     }
 
-    Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, body, { new: true }).then(todo => {
-      if (!todo) return res.status(404).json({ message: 'Todo not found' });
-      res.status(200).json({ todo, message: 'Todo was successfully updated' });
-    }).catch(e => res.status(400).send());
-  })
+    const todo = await Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, body, { new: true });
+    if (!todo) return res.status(404).json({ message: 'Todo not found' });
+    res.status(200).json({ todo, message: 'Todo was successfully updated' });
+  }))
   .get('/', (req, res) => {
     res.status(200).send('Hello there');
   })
-  .post('/users', (req, res) => {
+  .post('/users', asyncHandler(async (req, res) => {
     let body = _.pick(req.body, ['email', 'password']);
     let user = new User(body);
-    user.save()
-      .then(() => user.generateAuthToken())
-      .then(token => res.header('x-auth', token).status(201).send(user))
-      .catch(e => res.status(400).send({ error: e.message }));
-  })
+    await user.save();
+
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).status(201).send(user);
+  }))
   .get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
   })
-  .post('/users/login', (req, res) => {
+  .post('/users/login', asyncHandler(async (req, res) => {
     let body = _.pick(req.body, ['email', 'password']);
-    User.findByCredentials(body)
-      .then(user => {
-        user.generateAuthToken()
-          .then(token => res.header('x-auth', token).status(200).send({ user }));
-      })
-      .catch(e => res.status(400).send({ error: e }));
 
-  })
-  .delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token)
-      .then(() => {
-        res.status(200).send();
-      }, () => {
-        res.status(400).send();
-      });
+    const user = await User.findByCredentials(body);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).status(200).send({ user });
+  }))
+  .delete('/users/me/token', authenticate, asyncHandler(async (req, res) => {
+    await req.user.removeToken(req.token);
+    res.status(200).send();
+  }))
+  .use((err, req, res, next) => {
+    if (err.name === 'JsonWebTokenError') {
+      res.status(401).send({ error: err });
+    } else {
+      res.status(500).send(err.message);
+    }
+
   })
   .listen(port, () => console.log(`Server up and running at ${port}`));
